+++ kaskade/emojis.py
@@ -1,0 +1,3 @@
+THINKING_FACE = ":thinking_face:"
+FIRE = ":fire:"
+INFO = "[blue]:information:[/]"