+++ kaskade/styles.py
@@ -1,6 +1,7 @@
 from rich import box
 
 BORDER_FOCUSED = "green"
+BORDER_ERROR = "red"
 BORDER_MOUSE_OVER = "grey100"
 BORDER = "grey82"
 BOX = box.SQUARE