+++ kaskade/tui.py
@@ -1,5 +1,6 @@
 from typing import Optional, Type
 
+from confluent_kafka import KafkaException
 from rich.console import Console
 from textual.app import App
 from textual.driver import Driver
@@ -7,11 +8,13 @@ from textual.keys import Keys
 from textual.reactive import Reactive
 from textual.widget import Widget
 
+from kaskade import logger
 from kaskade.config import Config
 from kaskade.kafka.cluster_service import ClusterService
 from kaskade.kafka.models import Topic
 from kaskade.kafka.topic_service import TopicService
 from kaskade.utils.circular_list import CircularList
+from kaskade.widgets.error import Error
 from kaskade.widgets.footer import Footer
 from kaskade.widgets.header import Header
 from kaskade.widgets.help import Help
@@ -19,10 +22,13 @@ from kaskade.widgets.topic_detail import TopicDetail
 from kaskade.widgets.topic_header import TopicHeader
 from kaskade.widgets.topic_list import TopicList
 
+ERROR_MESSAGE_DELAY = 10
+
 
 class Tui(App):
     __topic: Optional[Topic] = None
     show_help = Reactive(False)
+    error = Reactive("")
 
     def __init__(
         self,
@@ -48,58 +54,86 @@ class Tui(App):
         self.cluster = self.cluster_service.current()
         self.topics = self.topic_service.list()
 
-        self.footer = Footer()
-        self.header = Header()
-        self.help = Help()
+        self.footer_widget = Footer()
+        self.header_widget = Header()
 
-        self.topic_list = TopicList()
-        self.topic_detail = TopicDetail()
-        self.topic_header = TopicHeader()
+        self.help_widget = Help()
+        self.error_widget = Error()
+
+        self.topic_list_widget = TopicList()
+        self.topic_detail_widget = TopicDetail()
+        self.topic_header_widget = TopicHeader()
         self.focusables = CircularList(
-            [self.topic_list, self.topic_header, self.topic_detail]
+            [self.topic_list_widget, self.topic_detail_widget]
         )
 
     async def on_mount(self) -> None:
-        self.help.layout_offset_x = -30
-        await self.view.dock(self.header, edge="top")
-        await self.view.dock(self.footer, edge="bottom")
-        await self.view.dock(self.topic_list, edge="left", size=40)
-        await self.view.dock(self.topic_header, self.topic_detail, edge="top")
-        await self.view.dock(self.help, edge="left", size=30, z=1)
+        self.help_widget.visible = False
+        self.error_widget.visible = False
+        await self.view.dock(self.header_widget, edge="top")
+        await self.view.dock(self.footer_widget, edge="bottom")
+        await self.view.dock(self.topic_list_widget, edge="left", size=40)
+        await self.view.dock(
+            self.topic_header_widget, self.topic_detail_widget, edge="top"
+        )
+        await self.view.dock(self.error_widget, edge="right", size=50, z=1)
+        await self.view.dock(self.help_widget, edge="right", size=30, z=1)
 
     async def on_load(self) -> None:
         await self.bind("q", "quit")
         await self.bind("Q", "quit")
         await self.bind("?", "toggle_help")
-        await self.bind(Keys.Escape, "default_view")
+        await self.bind(Keys.Escape, "back")
         await self.bind(Keys.F5, "reload_content")
         await self.bind(Keys.Left, "change_focus('{}')".format(Keys.Left))
         await self.bind(Keys.Right, "change_focus('{}')".format(Keys.Right))
 
+    async def watch_error(self, error: str) -> None:
+        show_error = bool(error)
+        if show_error:
+            self.focusables.reset()
+            await self.set_focus(self.error_widget)
+        self.error_widget.message = error
+        self.error_widget.visible = show_error
+
     async def watch_show_help(self, show_help: bool) -> None:
         if show_help:
-            self.help.layout_offset_x = self.view.size.width - 30
-        else:
-            self.help.layout_offset_x = -30
+            self.focusables.reset()
+            await self.set_focus(self.help_widget)
+        self.help_widget.visible = show_help
 
     async def action_toggle_help(self) -> None:
         self.show_help = not self.show_help
-        if self.show_help:
-            await self.set_focus(self.help)
-            self.focusables.reset()
 
-    async def action_default_view(self) -> None:
+    async def action_back(self) -> None:
         self.show_help = False
+        self.error = ""
+
+    def handle_exception(self, exception: Exception) -> None:
+        message = str(exception)
+
+        if isinstance(exception, KafkaException):
+            message = exception.args[0].str()
+
+        self.error = message
+
+        logger.critical("Error in runtime: %s", message)
+        logger.exception(exception)
 
     async def action_reload_content(self) -> None:
-        self.topics = self.topic_service.list()
+        try:
+            self.topics = self.topic_service.list()
+        except Exception as ex:
+            self.topics = []
+            self.handle_exception(ex)
+
         self.topic = None
         self.focusables.reset()
-        self.topic_list.scrollable_list = None
-        self.topic_detail.table = None
-        self.topic_list.refresh()
-        self.topic_header.refresh()
-        self.topic_detail.refresh()
+        self.topic_list_widget.scrollable_list = None
+        self.topic_detail_widget.table = None
+        self.topic_list_widget.refresh()
+        self.topic_header_widget.refresh()
+        self.topic_detail_widget.refresh()
         await self.set_focus(None)
 
     @property
@@ -109,10 +143,10 @@ class Tui(App):
     @topic.setter
     def topic(self, topic: Optional[Topic]) -> None:
         self.__topic = topic
-        self.topic_detail.table = None
-        self.topic_detail.tabs.index = 0
-        self.topic_detail.refresh()
-        self.topic_header.refresh()
+        self.topic_detail_widget.table = None
+        self.topic_detail_widget.tabs.index = 0
+        self.topic_detail_widget.refresh()
+        self.topic_header_widget.refresh()
 
     async def action_change_focus(self, key: Keys) -> None:
         focused: Widget = (