+++ kaskade/widgets/topic_header.py
@@ -2,7 +2,6 @@ from typing import Union
 
 from rich.align import Align
 from rich.panel import Panel
-from textual.reactive import Reactive
 from textual.widget import Widget
 
 from kaskade import styles
@@ -12,15 +11,6 @@ PANEL_SIZE = 5
 
 
 class TopicHeader(Widget):
-    has_focus: Reactive = Reactive(False)
-
-    def on_focus(self) -> None:
-        self.has_focus = True
-        self.app.focusables.current = self
-
-    def on_blur(self) -> None:
-        self.has_focus = False
-
     def on_mount(self) -> None:
         self.layout_size = PANEL_SIZE
 
@@ -35,7 +25,7 @@ class TopicHeader(Widget):
         panel = Panel(
             topic_info,
             title="topic",
-            border_style=styles.BORDER_FOCUSED if self.has_focus else styles.BORDER,
+            border_style=styles.BORDER,
             box=styles.BOX,
             title_align="left",
             padding=0,