+++ CHANGELOG.md
@@ -4,16 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Error message panel
+- Home folder
+- Command option info
+
 ## [0.2.9] - 2021-10-31
 ### Added
 - Release tags with prefix v
-
-[0.2.8]: https://github.com/sauljabin/kaskade/compare/0.2.7...0.2.8
 
 ## [0.2.8] - 2021-10-31
 ### Added
 - Testing release
 - Showing the messages count and lag
 
+[Unreleased]: https://github.com/sauljabin/kaskade/compare/v0.2.9...HEAD
 [0.2.9]: https://github.com/sauljabin/kaskade/compare/0.2.8...v0.2.9
 [0.2.8]: https://github.com/sauljabin/kaskade/compare/0.2.7...0.2.8